+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [unreleased]
 ### added
+* added `--no-visual-cues` flag to remove text-based graphics and visual cues
 * added status messages for all commands
   * commands should never output nothing
 * added help message for parsing unit format when value is missing or invalid